+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/InvertedIndexImpl.java
@@ -60,14 +60,14 @@ public class InvertedIndexImpl implements InvertedIndex {
     }
 
     @Override
-    public void insert(ResiduePairDescriptor residuePairDescriptor, Map<String, Collection<ResiduePairIdentifier>> residuePairOccurrences) {
+    public void insert(ResiduePairDescriptor residuePairDescriptor, Map<Integer, Collection<ResiduePairIdentifier>> residuePairOccurrences) {
         if (!paths) {
             ensureDirectoriesExist();
             this.paths = true;
         }
 
         try {
-            Map<Object, Object> data = residuePairOccurrences.entrySet()
+            Map<Integer, Object> data = residuePairOccurrences.entrySet()
                     .stream()
                     .collect(Collectors.toMap(Map.Entry::getKey,
                             entry -> entry.getValue()
@@ -75,16 +75,16 @@ public class InvertedIndexImpl implements InvertedIndex {
                                     .map(this::createObjectArray)
                                     .toArray()));
 
-            Map<Object, Object> map = getMap(residuePairDescriptor);
+            Map<Integer, Object> map = getMap(residuePairDescriptor);
 
             // read already present target identifiers and add to list to write
             data.putAll(map);
 
             // serialize message
-            ByteArrayOutputStream outputStream = MessagePack.encode(data);
-            Path path = getPath(residuePairDescriptor);
-
-            write(path, outputStream);
+            try (ByteArrayOutputStream outputStream = MessagePack.encode(data)) {
+                Path path = getPath(residuePairDescriptor);
+                write(path, outputStream);
+            }
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
@@ -94,14 +94,14 @@ public class InvertedIndexImpl implements InvertedIndex {
         try (OutputStream outputStream = Files.newOutputStream(path)) {
             OutputStream actual = gzipped ? new GZIPOutputStream(outputStream, BUFFER_SIZE) : outputStream;
             data.writeTo(actual);
+            actual.flush();
+            actual.close();
         }
     }
 
     @Override
-    public Stream<Pair<String, InvertedIndexResiduePairIdentifier[]>> select(ResiduePairDescriptor residuePairDescriptor) {
-        try {
-            InputStream inputStream = getInputStream(residuePairDescriptor);
-
+    public Stream<Pair<Integer, InvertedIndexResiduePairIdentifier[]>> select(ResiduePairDescriptor residuePairDescriptor) {
+        try (InputStream inputStream = getInputStream(residuePairDescriptor)) {
             // PSE can cause identifiers to flip - if so we need to flip them again to ensure correct overlap with other words
             return getPairs(inputStream, residuePairDescriptor);
         } catch (IOException e) {
@@ -109,10 +109,10 @@ public class InvertedIndexImpl implements InvertedIndex {
         }
     }
 
-    private Stream<Pair<String, InvertedIndexResiduePairIdentifier[]>> getPairs(InputStream inputStream, ResiduePairDescriptor residuePairDescriptor) throws IOException {
+    private Stream<Pair<Integer, InvertedIndexResiduePairIdentifier[]>> getPairs(InputStream inputStream, ResiduePairDescriptor residuePairDescriptor) throws IOException {
         return getData(inputStream)
                 .map(entry -> {
-                    String id = (String) entry.getKey();
+                    int id = entry.getKey();
                     Object[] array = (Object[]) entry.getValue();
                     InvertedIndexResiduePairIdentifier[] value = new InvertedIndexResiduePairIdentifier[array.length];
                     for (int i = 0; i < array.length; i++) {
@@ -127,8 +127,9 @@ public class InvertedIndexImpl implements InvertedIndex {
         return new InvertedIndexResiduePairIdentifier(data, residuePairDescriptor.isFlipped());
     }
 
-    private Stream<Map.Entry<Object, Object>> getData(InputStream inputStream) throws IOException {
-        return MessagePack.decode(inputStream).entrySet().stream();
+    @SuppressWarnings("unchecked")
+    private Stream<Map.Entry<Integer, Object>> getData(InputStream inputStream) throws IOException {
+        return ((Map<Integer, Object>) MessagePack.decode(inputStream)).entrySet().stream();
     }
 
     /**
@@ -149,16 +150,17 @@ public class InvertedIndexImpl implements InvertedIndex {
         return basePath.resolve(uberbin).resolve(bin + extension);
     }
 
-    private Map<Object, Object> getMap(ResiduePairDescriptor residuePairDescriptor) {
-        try {
-            return MessagePack.decode(getInputStream(residuePairDescriptor));
+    @SuppressWarnings("unchecked")
+    private Map<Integer, Object> getMap(ResiduePairDescriptor residuePairDescriptor) {
+        try (InputStream inputStream = getInputStream(residuePairDescriptor)) {
+            return (Map<Integer, Object>) MessagePack.decode(inputStream);
         } catch (IOException e) {
             return Collections.emptyMap();
         }
     }
 
     @Override
-    public void delete(Collection<String> removals) {
+    public void delete(Collection<Integer> removals) {
         if (!Files.exists(basePath)) {
             return;
         }
@@ -196,9 +198,9 @@ public class InvertedIndexImpl implements InvertedIndex {
         return new ResiduePairDescriptor(residueType1, residueType2, d1, d2, a);
     }
 
-    private void delete(ResiduePairDescriptor residuePairDescriptor, Collection<String> removals) {
+    private void delete(ResiduePairDescriptor residuePairDescriptor, Collection<Integer> removals) {
         try {
-            Map<Object, Object> map = getMap(residuePairDescriptor);
+            Map<Integer, Object> map = getMap(residuePairDescriptor);
 
             // if no entry would be removed: don't bother and return
             if (removals.stream().noneMatch(map::containsKey)) {
@@ -206,17 +208,17 @@ public class InvertedIndexImpl implements InvertedIndex {
             }
 
             // remove all occurrences of structure identifiers
-            Map<Object, Object> filteredMap = map.entrySet()
+            Map<Integer, Object> filteredMap = map.entrySet()
                     .stream()
                     // let only entries pass if their key is not in removal set
-                    .filter(entry -> !removals.contains((String) entry.getKey()))
+                    .filter(entry -> !removals.contains(entry.getKey()))
                     .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
 
             // serialize message
-            ByteArrayOutputStream outputStream = MessagePack.encode(filteredMap);
-            Path path = getPath(residuePairDescriptor);
-
-            write(path, outputStream);
+            try (ByteArrayOutputStream outputStream = MessagePack.encode(filteredMap)) {
+                Path path = getPath(residuePairDescriptor);
+                write(path, outputStream);
+            }
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }