+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/StructureIndexProvider.java
@@ -13,6 +13,20 @@ public interface StructureIndexProvider {
     String selectStructureIdentifier(int structureIndex);
 
     /**
+     * Maps between structure identifier and index.
+     * @param structureIdentifier the String
+     * @return the int
+     */
+    int selectStructureIndex(String structureIdentifier);
+
+    /**
+     * Get the structure index associated to this identifier. If none exists a new one will be created.
+     * @param structureIdentifier the structure identifier
+     * @return a unique int that maps to this identifier
+     */
+    int selectOrMintStructureIndex(String structureIdentifier);
+
+    /**
      * Get the next available structureIndex. This may reuse a previously assigned index that is no longer referenced.
      * @return an int that is guaranteed to uniquely identify a structure in the inverted index
      */